Rename component directory
