+++ client/app/components/OnboardingGuide/GuideCoverPhotoPage.js
