+++ client/app/components/OnboardingGuide/GuidePaypalPage.js
