+++ client/app/components/OnboardingGuide/GuideSloganAndDescriptionPage.js
