+++ client/app/components/OnboardingGuide/GuideStatusPage.js
