+++ client/app/components/OnboardingGuide/OnboardingGuideContainer.js
