+++ client/app/components/OnboardingGuide/styles.scss
